+++ core/reader/JsonLogFileReader.cpp
@@ -155,7 +155,6 @@ bool JsonLogFileReader::ParseLogLine(const char* buffer,
         return true;
     } else if (!mDiscardUnmatch) {
         LogParser::AddUnmatchLog(buffer, logGroup, logGroupSize);
-        return true;
     }
     return false;
 }